### 2017-03-07 #198 @whedon
[PRE REVIEW]: VCC: A framework for building containerized reproducible cluster software environments
**Submitting author:** @joshiggins (<a href="http://orcid.org/0000-0003-1551-5552">Joshua Higgins</a>)
**Repository:** <a href="https://github.com/hpchud/vccjs" target ="_blank">https://github.com/hpchud/vccjs</a>
**Version:** v1.0
