### 2017-03-07 2ba5123
make sure timeout for watcher binds this correctly
