### 2017-03-16 #208 @whedon
[REVIEW]: VCC: A framework for building containerized reproducible cluster software environments
**Submitting author:** @joshiggins (<a href="http://orcid.org/0000-0003-1551-5552">Joshua Higgins</a>)
**Repository:** <a href="https://github.com/hpchud/vccjs" target ="_blank">https://github.com/hpchud/vccjs</a>
**Version:** v1.0
**Archive:** <a href="http://dx.doi.org/10.6084/m9.figshare.4763857" target="_blank">10.6084/m9.figshare.4763857</a>
