### 2017-03-16 #208 @brainstorm
@joshiggins

1) The LICENSE seems a bit barebones to me... is it public domain? A clearer definition or reference would be great.

2) The update process fails on me due to hardcoded `/usr/bin` paths in the auto-update process. I would recommend to change the documentation and the `vcc` script so that it refers to `$HOME/bin` which has much higher chances to have suitable permissions rather than `/usr` and `/usr/local`:

```
$ vcc update
Updating VCC tool...

/usr/bin/vcc-new: Permission denied
There was a problem updating the script. Check proxy settings.
```

I'll continue reviewing, just wanted you to know early and as I go...
