### 2017-03-16 #208 @brainstorm
There are no performance claims to be confirmed in the `paper.md` file, or am I missing some benchmarks I have not seen?

Other than the points outlined above, all the rest looks great, good job @joshiggins!
