### 2017-03-16 #208 @pjotrp
Thanks @brainstorm for the fast turnaround. @joshiggins can you address the point? We also need to tick the OSI box for JOSS before acceptance.
