### 2017-03-16 #208 @pjotrp
On a personal note, it may be worth looking into Nix and Guix for true reproducibly installs (and they play well inside containers). Bruno has a nice [writeup](https://medium.com/@bmpvieira/a-truly-reproducible-scientific-paper-5059b282ee9a#.c7wny4wx6).
