### 2017-03-16 7375522
include a bit about scope in the readme
