### 2017-03-16 #208 @joshiggins
@brainstorm @pjotrp the MIT license text is copied from https://opensource.org/licenses/MIT as far as I know is the real deal?!

The claim that containers can be more performant than VMs is based on the reference _Felter, 2014_ (in `paper.bib`) but I wasnt sure how to indicate this in the markdown.

Thanks for checking everything! The `vcc` script - it will now try and update itself in whatever directory it's currently in, so will deal with `$HOME/bin` or `/usr/bin`, wherever, and give more helpful messages if this isn't possible.

I've bumped the release to `v1.1` for the update functionality (the released version of the script updates from a tag, not from `master`), hope this won't mess up the submission.....

p.s. the linked article was a good read! I will try those out for sure.
