### 2017-03-17 #208 @pjotrp
MIT license is OSI approved. @arfon I believe we can accept this paper using the latest release.
