### 2017-03-18 #208 @arfon
Hi @joshiggins - minor issue here - you actually need to cite the items in your `paper.bib` file for Pandoc to include them in the compiled paper PDF. You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html).
