### 2017-03-18 6d32672
cite items in paper.md
