### 2017-03-18 #208 @joshiggins
@arfon sorry about that, I missed the [@...] syntax in the example, but it is clear having looked again :laughing:
I just updated the `paper.md`.
