### 2017-03-18 #208 @arfon
@brainstorm many thanks for the review here and to @pjotrp for editing this one ✨

@joshiggins - your paper is now accepted into JOSS and your paper DOI is http://dx.doi.org/10.21105/joss.00208 ⚡️ 🚀 💥
